requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.9
numpy>=1.24
opencv-python>=4.8
pillow>=10.0
pyyaml>=6.0
tqdm>=4.66
tensorboard>=2.14
albumentations>=1.4.0
scikit-learn>=1.3
